just create another store with a different config rather than having two store files
