--- examples/counter-react/src/store/__tests__/counterStoreAsIs.store.spec.js
@@ -1,37 +1,0 @@
-import React, { useContext } from 'react';
-import CounterStoreAsIs from '../counterStoreAsIs.store';
-import { render } from '@testing-library/react';
-import { act } from 'react-dom/test-utils';
-
-describe('CounterStore', () => {
-  let store;
-  const renderStore = () => {
-    let Prep = CounterStoreAsIs.Provider(() => {
-      store = useContext(CounterStoreAsIs.Context);
-      return null;
-    });
-    render(<Prep />);
-  };
-
-  it('tests increment', () => {
-    act(() => {
-      renderStore();
-    });
-    expect(store.state.count).toEqual(0);
-    act(() => {
-      store.incrementCount();
-    });
-    expect(store.state.count).toEqual(1);
-  });
-
-  it('tests decrement', () => {
-    act(() => {
-      renderStore();
-    });
-    expect(store.state.count).toEqual(0);
-    act(() => {
-      store.decrementCount();
-    });
-    expect(store.state.count).toEqual(-1);
-  });
-});
