+++ examples/counter-react/src/store/index.js
@@ -1,3 +1,2 @@
-export { default as CounterStore } from './counter.store';
+export { CounterStore, CounterStoreAsIs } from './counter.store';
 export { default as CounterWithReducerStore } from './counterWithReducer.store.js';
-export { default as CounterStoreAsIs } from './counterStoreAsIs.store';
